Improve HT hashing function

+++ src/ds/ph_hashtable.h
@@ -36,7 +36,7 @@
 typedef struct _ph_bucket_t {
     ph_string_t *key; // @todo remove pointer to key?
     void *value;
-    int hash;
+    long hash;
     int variance;
 } ph_bucket_t;
 
@@ -51,14 +51,14 @@ typedef struct _ph_hashtable_t {
 
 void ph_hashtable_init(ph_hashtable_t *ht, int size, void (*dtor)(void *));
 void ph_hashtable_insert(ph_hashtable_t *ht, ph_string_t *key, void *value);
-void ph_hashtable_insert_ind(ph_hashtable_t *ht, int hash, void *value);
+void ph_hashtable_insert_ind(ph_hashtable_t *ht, long hash, void *value);
 void ph_hashtable_delete(ph_hashtable_t *ht, ph_string_t *key);
-void ph_hashtable_delete_ind(ph_hashtable_t *ht, int hash);
+void ph_hashtable_delete_ind(ph_hashtable_t *ht, long hash);
 void *ph_hashtable_search(ph_hashtable_t *ht, ph_string_t *key);
-void *ph_hashtable_search_ind(ph_hashtable_t *ht, int hash);
+void *ph_hashtable_search_ind(ph_hashtable_t *ht, long hash);
 ph_string_t *ph_hashtable_key_fetch(ph_hashtable_t *ht, ph_string_t *key);
 void ph_hashtable_update(ph_hashtable_t *ht, ph_string_t *key, void *value);
-void ph_hashtable_update_ind(ph_hashtable_t *ht, int hash, void *value);
+void ph_hashtable_update_ind(ph_hashtable_t *ht, long hash, void *value);
 void ph_hashtable_destroy(ph_hashtable_t *ht);
 void ph_hashtable_to_hashtable(HashTable *ht, ph_hashtable_t *phht);
 void *ph_hashtable_random_value(ph_hashtable_t *ht);
